adapters: split postgres support owners
